backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-jose[cryptography]
passlib[bcrypt]
google-generativeai>=0.3.04
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
bcrypt==4.0.1
